backend/requirements.txt
aiofiles==0.7.0
cachetools==4.2.2
certifi==2021.5.30
click==8.0.1
fastapi==0.65.3
google-api-core==1.30.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.20.0
google-cloud-bigquery-storage==2.5.0
google-cloud-core==1.7.1
google-cloud-storage==1.40.0
gunicorn==20.1.0
Jinja2==3.0.1
joblib==1.0.1
matplotlib==3.4.2
numpy==1.21.0
openai==0.9.3
openpyxl==3.0.7
pandas==1.3.0
pandas-gbq==0.15.0
Pillow==8.3.1
plotly==5.1.0
pyarrow==4.0.1
pydantic==1.8.2
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.20
starlette==0.14.2
urllib3==1.26.6
uvicorn==0.14.0
websockets==9.1
celery==5.1.2
redis
flower